fix: add basedpyright static type checking, fix real bugs it surfaced
Fase 0 of the enterprise roadmap: closing out the foundation before
scaling. Added basedpyright to the backend (previously no static type
checking existed) and wired it into CI as a blocking step. Went from
411 raw errors to 0 by fixing real bugs and, separately, silencing
confirmed stub-precision false positives with documented reasons.

Real bugs fixed:
- 20 call sites across ai_report.py, ai_screener.py, autopilot.py, and
  data_fetcher.py checked `isinstance(result, Exception)` on the output
  of `asyncio.gather(..., return_exceptions=True)`. asyncio.CancelledError
  is a BaseException, not an Exception — under cancellation (a client
  disconnect, a timeout) it would leak through as if it were valid data,
  and any code calling `.get()` on it would crash. Broadened to
  `BaseException` everywhere this pattern appears.
- binance_service.py: change_leverage() posted a raw urlencoded string
  via httpx's `data=` param with no Content-Type header — switched to
  `content=` with an explicit `application/x-www-form-urlencoded`
  header, matching Binance's documented signed-request format.
- 3 `sum()` calls over Decimal generators in paper_trades.py (fees,
  cash, debit) had no `start=`, so an empty fills list would silently
  produce the literal int 0 instead of Decimal(0) — and a later
  `.quantize()` call would then crash with AttributeError.
- market_sources.py: SourceError.__init__ never called
  super().__init__(), so str(exc) printed the raw args tuple instead of
  the message in every log line that formats the exception.
- evt.py / copulas.py: several `simulate()`/`return_level()` methods
  used post-fit() attributes (xi, sigma, theta, corr_matrix, df)
  without checking they were set, so calling them before fit() crashed
  with a confusing TypeError instead of a clear "not fitted" error.
  Added the guard consistently across GPDModel, StudentT/Clayton/Gumbel/
  Frank copulas.
- ai_factory.py / options_api.py: Claude's response content block was
  read via `.text` unconditionally; on any non-text block (tool use,
  thinking) this would crash. Added an isinstance(TextBlock) check.
  Same class of gap fixed for the new GeminiProvider (response.text can
  be None when content is blocked by safety filters).
- ai_report.py: aggregate_signals() could receive None if the "agents"
  gather task resolved to None instead of a list.

Everything else was scoped, documented stub-precision noise (scipy's
linregress/minimize_scalar/brentq, pandas DataFrame.groupby/.get(),
redis-py's dual sync/async ResponseT, openpyxl's `Workbook.active`) —
each left as a `# pyright: ignore[...]` with a one-line reason, not a
blanket rule suppression.

+++ backend/app/models/black_swan.py
@@ -193,7 +193,7 @@ class BlackSwanDetector:
                 "negative_keywords_found": neg_count,
             })
 
-        avg_sentiment = BlackSwanDetector._finite_float(np.mean(sentiment_scores)) if sentiment_scores else 0.0
+        avg_sentiment = BlackSwanDetector._finite_float(float(np.mean(sentiment_scores))) if sentiment_scores else 0.0
         alert_level = "CRITICAL" if avg_sentiment < -0.5 else (
             "WARNING" if avg_sentiment < -0.2 else (
                 "WATCH" if avg_sentiment < 0 else "NORMAL"
